# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DMS - Document Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: dossierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: assignedToMe
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Archive/delete document (logical)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/status:
    patch:
      operationId: updateDocumentStatus
      tags: [documents]
      summary: Update document status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentStatusUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/versions:
    get:
      operationId: getDocumentVersions
      tags: [documents]
      summary: Get document version history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"
  /documents/{id}/assignments:
    post:
      operationId: assignDocument
      tags: [documents]
      summary: Assign document to users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "200":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # --- PROTOCOLS ---
...
  /dossiers:
    get:
      operationId: listDossiers
      tags: [dossiers]
      summary: List dossiers
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: responsibleId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Dossier list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DossierList"
    post:
      operationId: createDossier
      tags: [dossiers]
      summary: Create a new dossier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DossierInput"
      responses:
        "201":
          description: Created dossier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dossier"
  /dossiers/{id}:
    get:
      operationId: getDossier
      tags: [dossiers]
      summary: Get dossier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dossier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dossier"
    patch:
      operationId: updateDossier
      tags: [dossiers]
      summary: Update dossier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DossierUpdate"
      responses:
        "200":
          description: Updated dossier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dossier"
  /dossiers/{id}/documents:
    get:
      operationId: getDossierDocuments
      tags: [dossiers]
      summary: List documents in a dossier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Documents in dossier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  # --- TASKS ---
...
components:
  schemas:
...
    DocumentList:
      type: object
      required:
        - items
        - total
        - page
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Document:
      type: object
      required:
        - id
        - title
        - type
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        status:
          type: string
        subject:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        confidentiality:
          type: string
          default: normal
        priority:
          type: string
          default: normal
        version:
          type: integer
          default: 1
        driveUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        dossierId:
          type: ["integer", "null"]
        dossierTitle:
          type: ["string", "null"]
        protocolId:
          type: ["integer", "null"]
        protocolNumber:
          type: ["string", "null"]
        classificationId:
          type: ["integer", "null"]
        classificationCode:
          type: ["string", "null"]
        responsibleId:
          type: ["integer", "null"]
        responsibleName:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: string
        tags:
          type: array
          items:
            type: string
        ocrText:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        archivedAt:
          type: ["string", "null"]
    DocumentInput:
      type: object
      required:
        - title
        - type
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
        subject:
          type: string
        description:
          type: string
        confidentiality:
          type: string
        priority:
          type: string
        dossierId:
          type: ["integer", "null"]
        classificationId:
          type: ["integer", "null"]
        responsibleId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        driveUrl:
          type: string
        fileName:
          type: string
    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        description:
          type: string
        type:
          type: string
        confidentiality:
          type: string
        priority:
          type: string
        dossierId:
          type: ["integer", "null"]
        classificationId:
          type: ["integer", "null"]
        responsibleId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        aiSummary:
          type: string
    DocumentStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
        note:
          type: string
    DocumentVersion:
      type: object
      required:
        - id
        - documentId
        - version
        - status
        - createdAt
        - createdById
        - createdByName
      properties:
        id:
          type: integer
        documentId:
          type: integer
        version:
          type: integer
        status:
          type: string
        description:
          type: ["string", "null"]
        driveUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        fileHash:
          type: ["string", "null"]
        createdAt:
          type: string
        createdById:
          type: integer
        createdByName:
          type: string
    AssignmentInput:
      type: object
      required:
        - assignments
      properties:
        assignments:
          type: array
          items:
            type: object
            required:
              - userId
              - role
            properties:
              userId:
                type: integer
              role:
                type: string
        note:
          type: string

    # --- PROTOCOL ---
...
    DossierList:
      type: object
      required:
        - items
        - total
        - page
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Dossier"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Dossier:
      type: object
      required:
        - id
        - code
        - title
        - status
        - year
        - openedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        code:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        year:
          type: integer
        area:
          type: ["string", "null"]
        confidentiality:
          type: string
          default: normal
        responsibleId:
          type: ["integer", "null"]
        responsibleName:
          type: ["string", "null"]
        classificationId:
          type: ["integer", "null"]
        classificationCode:
          type: ["string", "null"]
        documentCount:
          type: integer
          default: 0
        protocolCount:
          type: integer
          default: 0
        openedAt:
          type: string
        closedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DossierInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        area:
          type: string
        confidentiality:
          type: string
        responsibleId:
          type: ["integer", "null"]
        classificationId:
          type: ["integer", "null"]
    DossierUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        area:
          type: string
        confidentiality:
          type: string
        responsibleId:
          type: ["integer", "null"]
        classificationId:
          type: ["integer", "null"]
        closedAt:
          type: ["string", "null"]

    # --- TASKS ---
...