# lib/api-zod/src/generated/types/documentList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DMS - Document Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';

export interface DocumentList {
  items: Document[];
  total: number;
  page: number;
  limit: number;
}